chore(map): move old header
